backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.28.0
aiohttp==3.8.6
feedparser==6.0.10
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pytest-asyncio==0.21.1
pytest==7.4.3
playwright==1.42.0
beautifulsoup4==4.12.3
celery[redis]==5.3.6